+++ NEWS.md
@@ -1,3 +1,10 @@
+# ggtangle 0.0.4.001
+
++ 'node_label' in `cnetplot()` supports new arguments (2024-11-05, Tue)
+    - 'exclusive' to label genes that is uniquely belong to categories
+    - 'share' to label genes that are shared between categories
+    - expresson, like '> 1' or '< 1', to label those genes with `foldChange > 1` or `foldChange < 1`
+
 # ggtangle 0.0.4
 
 + compatible with more igraph layouts (2024-10-29, Tue)